pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mexam"
authors = [{name = "Oliver Lindemann", email = "lindemann@essb.eur.nl"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
requires-python = ">=3.10"
dependencies =  ["PyYAML>=5.2",  "toml>=0.10", "typing_extensions>=4.12"]

[project.scripts]
mexam = "mexam.__main__:run"
